+++ app/include_upgrade_testnet.go
@@ -12,7 +12,7 @@ import (
 // it is also used for e2e testing
 func init() {
 	Upgrades = []upgrades.Upgrade{v1.CreateUpgrade(v1.UpgradeDataString{
-		BtcStakingParamStr:        testnet.BtcStakingParamStr,
+		BtcStakingParamsStr:       testnet.BtcStakingParamsStr,
 		FinalityParamStr:          testnet.FinalityParamStr,
 		IncentiveParamStr:         testnet.IncentiveParamStr,
 		CosmWasmParamStr:          testnet.CosmWasmParamStr,