+++ test/e2e/configurer/chain/queries_btcstaking.go
@@ -23,6 +23,18 @@ func (n *NodeConfig) QueryBTCStakingParams() *bstypes.Params {
 	return &resp.Params
 }
 
+func (n *NodeConfig) QueryBTCStakingParamsByVersion(version uint32) *bstypes.Params {
+	path := fmt.Sprintf("/babylon/btcstaking/v1/params/%d", version)
+	bz, err := n.QueryGRPCGateway(path, url.Values{})
+	require.NoError(n.t, err)
+
+	var resp bstypes.QueryParamsByVersionResponse
+	err = util.Cdc.UnmarshalJSON(bz, &resp)
+	require.NoError(n.t, err)
+
+	return &resp.Params
+}
+
 func (n *NodeConfig) QueryFinalityParams() *ftypes.Params {
 	bz, err := n.QueryGRPCGateway("/babylon/finality/v1/params", url.Values{})
 	require.NoError(n.t, err)