+++ testutil/btcstaking-helper/keeper.go
@@ -168,6 +168,7 @@ func (h *Helper) GenAndApplyCustomParams(
 		MinUnbondingTimeBlocks:    minUnbondingTime,
 		UnbondingFeeSat:           1000,
 		AllowListExpirationHeight: allowListExpirationHeight,
+		BtcActivationHeight:       1,
 	})
 	h.NoError(err)
 	return covenantSKs, covenantPKs
@@ -218,6 +219,26 @@ func (h *Helper) CreateDelegation(
 	usePreApproval bool,
 	addToAllowList bool,
 ) (string, *types.MsgCreateBTCDelegation, *types.BTCDelegation, *btclctypes.BTCHeaderInfo, *types.InclusionProof, *UnbondingTxInfo, error) {
+	return h.CreateDelegationWithBtcBlockHeight(
+		r, delSK, fpPK, changeAddress, stakingValue,
+		stakingTime, unbondingValue, unbondingTime,
+		usePreApproval, addToAllowList, 10,
+	)
+}
+
+func (h *Helper) CreateDelegationWithBtcBlockHeight(
+	r *rand.Rand,
+	delSK *btcec.PrivateKey,
+	fpPK *btcec.PublicKey,
+	changeAddress string,
+	stakingValue int64,
+	stakingTime uint16,
+	unbondingValue int64,
+	unbondingTime uint16,
+	usePreApproval bool,
+	addToAllowList bool,
+	btcBlockHeight uint32,
+) (string, *types.MsgCreateBTCDelegation, *types.BTCDelegation, *btclctypes.BTCHeaderInfo, *types.InclusionProof, *UnbondingTxInfo, error) {
 	stakingTimeBlocks := stakingTime
 	bsParams := h.BTCStakingKeeper.GetParams(h.Ctx)
 	bcParams := h.BTCCheckpointKeeper.GetParams(h.Ctx)
@@ -261,7 +282,7 @@ func (h *Helper) CreateDelegation(
 	prevBlock, _ := datagen.GenRandomBtcdBlock(r, 0, nil)
 	btcHeaderWithProof := datagen.CreateBlockWithTransaction(r, &prevBlock.Header, testStakingInfo.StakingTx)
 	btcHeader := btcHeaderWithProof.HeaderBytes
-	btcHeaderInfo := &btclctypes.BTCHeaderInfo{Header: &btcHeader, Height: 10}
+	btcHeaderInfo := &btclctypes.BTCHeaderInfo{Header: &btcHeader, Height: btcBlockHeight}
 	serializedStakingTx, err := bbn.SerializeBTCTx(testStakingInfo.StakingTx)
 	h.NoError(err)