+++ x/btcstaking/keeper/params_test.go
@@ -1,6 +1,7 @@
 package keeper_test
 
 import (
+	"fmt"
 	"math"
 	"math/rand"
 	"testing"
@@ -16,10 +17,28 @@ func TestGetParams(t *testing.T) {
 	k, ctx := testkeeper.BTCStakingKeeper(t, nil, nil, nil)
 	params := types.DefaultParams()
 
+	currParams := k.GetParams(ctx)
+	params.BtcActivationHeight = currParams.BtcActivationHeight + 1
 	err := k.SetParams(ctx, params)
 	require.NoError(t, err)
 
 	require.EqualValues(t, params, k.GetParams(ctx))
+}
+
+func TestAddNewPairParams(t *testing.T) {
+	htvm := types.NewHeightToVersionMap()
+	// btc start height, version of params
+	err := htvm.AddNewPair(10, 0)
+	require.NoError(t, err)
+
+	err = htvm.AddNewPair(11, 1)
+	require.NoError(t, err)
+
+	err = htvm.AddNewPair(11, 2)
+	require.EqualError(t, err, fmt.Errorf("pairs must be sorted by start height in ascending order, got %d <= %d", 11, 11).Error())
+
+	err = htvm.AddNewPair(15, 1)
+	require.EqualError(t, err, fmt.Errorf("versions must be strictly increasing, got %d != %d + 1", 1, 1).Error())
 }
 
 func TestGetParamsVersions(t *testing.T) {
@@ -33,6 +52,7 @@ func TestGetParamsVersions(t *testing.T) {
 
 	params1 := types.DefaultParams()
 	params1.MinSlashingTxFeeSat = 23400
+	params1.BtcActivationHeight = pv.Params.BtcActivationHeight + 1
 
 	err := k.SetParams(ctx, params1)
 	require.NoError(t, err)
@@ -62,14 +82,17 @@ func FuzzParamsVersioning(f *testing.F) {
 		var generatedParams []*types.Params
 		generatedParams = append(generatedParams, &params0)
 
+		var btcActivationHeights []uint32
 		for i := 0; i < numVersionsToGenerate; i++ {
 			params := types.DefaultParams()
 			// randomize two parameters so each params are slightly different
 			params.MinSlashingTxFeeSat = r.Int63()
 			params.MinUnbondingTimeBlocks = uint32(r.Intn(math.MaxUint16))
+			params.BtcActivationHeight = uint32(i) + 1
 			err := k.SetParams(ctx, params)
 			require.NoError(t, err)
 			generatedParams = append(generatedParams, &params)
+			btcActivationHeights = append(btcActivationHeights, params.BtcActivationHeight)
 		}
 
 		allParams := k.GetAllParams(ctx)
@@ -90,5 +113,19 @@ func FuzzParamsVersioning(f *testing.F) {
 		lastVer := k.GetParamsByVersion(ctx, uint32(len(generatedParams)-1))
 		require.EqualValues(t, *generatedParams[len(generatedParams)-1], lastParams)
 		require.EqualValues(t, lastParams, *lastVer)
+
+		heightToVersionMap := k.GetHeightToVersionMap(ctx)
+		require.NotNil(t, heightToVersionMap)
+		require.EqualValues(t, len(generatedParams), len(heightToVersionMap.Pairs))
+
+		for _, btcActivationHeight := range btcActivationHeights {
+			paramsBTCActivation, version, err := k.GetParamsForBtcHeight(ctx, uint64(btcActivationHeight))
+			require.NoError(t, err)
+			require.NotNil(t, paramsBTCActivation)
+
+			paramsByVersion := k.GetParamsByVersion(ctx, version)
+			require.NotNil(t, paramsByVersion)
+			require.EqualValues(t, *paramsBTCActivation, *paramsByVersion)
+		}
 	})
 }